requirements.txt
pytube~=15.0.0
python-dotenv~=1.0.1
spotipy~=2.24.0